+++ app/Http/Livewire/Admin/AdminBrandComponent.php
@@ -4,19 +4,22 @@ namespace App\Http\Livewire\Admin;
 
 use App\Models\Brand;
 use Livewire\Component;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\WithPagination;
 use Illuminate\Support\Str;
 class AdminBrandComponent extends Component
 {
     use WithPagination;
-
+    use AuthorizesRequests;
     public $name;
     public $slug;
-    protected $brand_id;
+    public $brand_id;
     public $showEditModal = False;
+
 
     //delete brand
     public function deleteBrand($brand_id){
+        $this->authorize('brand-delete');
         $brand = Brand::find($brand_id);
         $brand->delete();
         session()->flash('message','Xoá thương hiệu thành công');
@@ -24,7 +27,9 @@ class AdminBrandComponent extends Component
 
     //refesh page
     public function refesh(){
-        $this->emitTo('admin.admin.admin-brand-component','refreshComponent');
+        $this->emitTo('admin.admin-brand-component','refreshComponent');
+        $this->emitTo('admin.admin-add-brand-component','refreshComponent');
+        $this->emitTo('footer-component','refreshComponent');
         $this->dispatchBrowserEvent('hide-form');
     }
 
@@ -52,6 +57,7 @@ class AdminBrandComponent extends Component
         ]);
     }
     public function updateBrand(){
+        $this->authorize('brand-edit');
         $this->validate([
             'name' => 'required',
             'slug' => 'required|unique:brands'
@@ -69,7 +75,7 @@ class AdminBrandComponent extends Component
 
     public function render()
     {
-        $brands = Brand::paginate(12);
+        $brands = Brand::all();
         return view('livewire.admin.admin-brand-component',['brands'=>$brands])->layout('layouts.base');
     }
 }
