+++ app/Http/Livewire/Admin/AdminHomeSliderComponent.php
@@ -4,10 +4,12 @@ namespace App\Http\Livewire\Admin;
 
 use App\Models\HomeSlider;
 use Livewire\Component;
-
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 class AdminHomeSliderComponent extends Component
 {
+    use AuthorizesRequests;
     public function deleteSlider($id){
+        $this->authorize('slider-delete');
         $slider = HomeSlider::find($id);
         if($slider->image){
             unlink('assets/images/sliders'.'/'.$slider->image);
